remove broken send function
There is no way to send back to the faucet using a method
on the faucet contract, even if we use `approve`.  This is
because the message sender has to be the owner, but the token
contract will think the faucet is the message sender.

So the way to setup "send back to faucet" functionality
is to have the GUI interact directly with the token contract,
sending from the user to the faucet address.